wait_for_firing_alert_sampler(..) to honor timeout (#324)
The method monitoring.wait_for_firing_alert_sampler doesn't
honor timeout, when the timeout value is a different value
other than 10 minutes. Because of this behavior, few
observability test cases that makes use of this method with
custom timeout values like 15 mins fails as the method could
only wait for 10mins max.

Now this issue is fixed by passing the custom timeout value
to the other method that it calls.